src/requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via httpx
boto3==1.38.35
    # via
    #   madeinabyss-faiss-search (pyproject.toml)
    #   langchain-aws
botocore==1.38.35
    # via
    #   boto3
    #   s3transfer
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
faiss-cpu==1.9.0
    # via madeinabyss-faiss-search (pyproject.toml)
h11==0.16.0
    # via httpcore
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via langsmith
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.25
    # via madeinabyss-faiss-search (pyproject.toml)
langchain-aws==0.2.25
    # via madeinabyss-faiss-search (pyproject.toml)
langchain-core==0.3.65
    # via
    #   langchain
    #   langchain-aws
    #   langchain-text-splitters
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.3.45
    # via
    #   langchain
    #   langchain-core
numpy==2.2.6
    # via
    #   madeinabyss-faiss-search (pyproject.toml)
    #   faiss-cpu
    #   langchain-aws
orjson==3.10.18
    # via langsmith
packaging==24.2
    # via
    #   faiss-cpu
    #   langchain-core
    #   langsmith
pydantic==2.11.5
    # via
    #   langchain
    #   langchain-aws
    #   langchain-core
    #   langsmith
pydantic-core==2.33.2
    # via pydantic
python-dateutil==2.9.0.post0
    # via botocore
pyyaml==6.0.2
    # via
    #   langchain
    #   langchain-core
requests==2.32.4
    # via
    #   langchain
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
s3transfer==0.13.0
    # via boto3
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via langchain
tenacity==9.1.2
    # via langchain-core
typing-extensions==4.14.0
    # via
    #   anyio
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via
    #   botocore
    #   requests
xmltodict==0.14.2
    # via madeinabyss-faiss-search (pyproject.toml)
zstandard==0.23.0
    # via langsmith
